refactor(web): move Database out of Cortex into a standalone top-level page
The Database tool operates data, not knowledge — it doesn't belong under
the Cortex (memory/notes/knowledge) project workspace. Promote it to a
first-class sidebar entry (/database, shortcut g d) that picks a project
then shows the same connections/schema/grid/SQL panel. Connections stay
per-project (cwd-scoped), so the REST/MCP backend is unchanged. Extract
the cwd picker into a shared ProjectCwdPicker reused by /cortex/project
and /database; drop the Database tab from ProjectScreen. i18n parity green.
