+++ app/web/src/router.tsx
@@ -11,6 +11,7 @@ import { SessionsPage } from '@/pages/Sessions'
 import { ProvidersPage } from '@/pages/Providers'
 import { ChannelsPage } from '@/pages/Channels'
 import { IntegrationsPage } from '@/pages/Integrations'
+import { DatabasePage } from '@/pages/Database'
 import { ActivityPage } from '@/pages/Activity'
 import { MemoryPage } from '@/pages/Memory'
 import { MemoryWorkersPage } from '@/pages/MemoryWorkers'
@@ -81,6 +82,15 @@ const integrationsRoute = createRoute({
   getParentRoute: () => protectedRoute,
   path: '/integrations',
   component: IntegrationsPage,
+})
+
+const databaseRoute = createRoute({
+  getParentRoute: () => protectedRoute,
+  path: '/database',
+  component: DatabasePage,
+  validateSearch: (search) => ({
+    cwd: typeof search.cwd === 'string' ? search.cwd : '',
+  }),
 })
 
 const activityRoute = createRoute({
@@ -250,6 +260,7 @@ const routeTree = rootRoute.addChildren([
     providersRoute,
     channelsRoute,
     integrationsRoute,
+    databaseRoute,
     activityRoute,
     cortexRoute,
     cortexProjectRoute,
